refactor: move utility functions and shared state to lib directory; enhance organization and maintainability
