+++ packages/@studiocms/wysiwyg/src/lib/shared.ts
